Move remaining Sign() tests to ubirch_sim_test.go
